requirements.txt
numpy
scipy
matplotlib
opencv-python>=3.4.3.18
tensorflow>=1.14
keras
wget
tqdm
patool